Implemented Discord Notifications
Implemented code to send a POST request to discord with the specified discord channel ID's that should be notified when a YouTube video is uploaded.

The youtube_notification class will check if a notification has already been sent for the given video ID. (On occasion, YouTube will send multiple webhooks to the callback for a single video)

A formatted embed message is sent to Discord.