Updates Discord Module to Reflect API Changes
Updated the discord commands that interface with the API to properly expect and handle the new JSON formatted data. Checking for the status code is not particularly important as the value in "message" is designed to inform the user of success or failure. In the event that "message" cannot be found in the response content, a generalized "Back-end server error!" will be returned.