Implemented Display Name Updating for YouTube Channels
Implemented a method of updating the YouTube display name field in the database ('youtube_channels'). This value is used for creating a list of notifications in route '/notifications' and command '!list'. In addition to being updated routinely, the YouTube display name will be added to the 'youtube_channels' table when a new channel is added to the DB.
This update is triggered via a PUT request to the API: '/youtube/update'

The "cron" task no longer handles the code for updating the YouTube tables and webhook subscriptions. It instead creates a PUT requests to the API route '/youtube/update'. This begins to eliminate some of the current segmentation between the "cron" and "API" modules.

Additionally, added a 30 second sleep to the beginning of “cron_run.py”. This was added to ensure that the API has had adequate time to start.