Updated get_youtube_notifications
Updated the function: "get_youtube_notifications()" in class "notifications”

These changes reflect the addition of the "yt_display_name" field added to the "youtube_channels” table. The SQL queries were optimized and reduced to a single query that joins the "yt_display_name" to the rows selected in "youtube_notifications". This eliminates the need to flatten the initial query into a list using the lambda function.

Also updated the list command to show the channel display name with a hyperlink to the channel instead of just a link.