Implemented Notification Limits with Overrides
Implemented a method of limiting the number of notifications on a per guild basis. In addition, the bot owner can increase the notification limit for each guild. When a guild reaches the notification limit they'll be presented with a message indicating how to manage existing notifications. A command has been created to allow guild owners/admins to request a limit increase. Currently, the limit increase is done manually.
New commands:
-	k!request -  Notifies the developer that a guild has requested a limit increase. Also provides guild information such as id, name, and member_count.
-	k!override – This command can only be used by the bot owner. It is used to check, set, and delete notification limit overrides.
API Routes:
-	GET /notifications/limits -  Returns the current limits for the given Discord guild ID.
-	Put /notifications/limits – Creates/updates an override.
-	DELETE /notifications/limits – Deletes an override.
Environment Variables:
-	TWITCH_NOTIFICATION_LIMIT – Default notification limit for Twitch.
-	YOUTUBE_NOTIFICATION_LIMIT – Default notification limit for YouTube.